+++ js/usecases/servicos/criar-servico-usecase.js
@@ -1,0 +1,13 @@
+const CriarServicoUseCase = {
+    async executar(dados = {}, service = ServicoService) {
+        try {
+            return service.criarServico(dados);
+        } catch (erro) {
+            return {
+                sucesso: false,
+                servico: null,
+                erros: [erro.message || "Erro ao criar Servico."]
+            };
+        }
+    }
+};